Split fix.js to several files
